addressing nits and renaming gocse to go
