+++ README.md
@@ -12,4 +12,4 @@ For the Node.js implementation, see the `nodejs` directory.
 
 For the C# implementation, see the `dotnet` directory.
 
-For the Go implementation, see the `gocse` directory.
+For the Go implementation, see the `go` directory.
